Move Serve method to api file
